refactor(consensus): promote committer tests to integration tests
Move base_committer_tests, multi_committer_tests, and
pipelined_committer_tests from crates/consensus/src/tests/ (the
Rust-2018 sub-module pattern) to crates/consensus/tests/ as proper
integration tests over the public crate API. Required visibility
bumps:

- `LeaderElector` and its `new` / `elect_leader` methods: pub(crate) -> pub.
  Tests assert which authority is the leader at a given round; this is
  a legitimate API surface, and per project_leader_election future
  policies will want a public abstraction here anyway.

- `Committer::try_commit` (inherent): pub(crate) -> pub. The trait-method
  version on `DagConsensus` is already public; the inherent method had
  an extra `+ '_` lifetime that the inherent method offers but the
  trait method does not, and tests bind through the inherent path.

multi_committer_tests and pipelined_committer_tests will be deleted in
#138 when the per-protocol macro suite subsumes them; keeping them here
avoids a coverage gap for leader_count > 1 and pipeline = true
scenarios in the interim.

Sets up #135 -> #140.
