+++ crates/consensus/src/leader.rs
@@ -6,17 +6,17 @@ use dag::{authority::Authority, block::RoundNumber};
 /// Determines the leader for each round. Different
 /// consensus protocols may use different election
 /// strategies; this struct encapsulates that choice.
-pub(crate) struct LeaderElector {
+pub struct LeaderElector {
     committee_len: usize,
 }
 
 impl LeaderElector {
-    pub(crate) fn new(committee_len: usize) -> Self {
+    pub fn new(committee_len: usize) -> Self {
         Self { committee_len }
     }
 
     /// Round-robin leader election.
-    pub(crate) fn elect_leader(&self, round: RoundNumber) -> Authority {
+    pub fn elect_leader(&self, round: RoundNumber) -> Authority {
         Authority::new(round % self.committee_len as u64)
     }
 }
