--- crates/consensus/src/tests/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-mod base_committer_tests;
-mod multi_committer_tests;
-mod pipelined_committer_tests;
