+++ crates/consensus/tests/base_committer_tests.rs
@@ -3,7 +3,7 @@
 
 use std::num::NonZeroUsize;
 
-use crate::{committer::Committer, leader::LeaderElector, protocol::Protocol};
+use consensus::{committer::Committer, leader::LeaderElector, protocol::Protocol};
 use dag::{
     authority::Authority,
     block::RoundNumber,
@@ -13,50 +13,44 @@ use dag::{
     test_util::{build_dag, build_dag_layer, committee},
 };
 
-/// Commit the leaders of the first wave.
+/// Commit one leader.
 #[test]
 #[tracing_test::traced_test]
 fn direct_commit() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
-    let wave_length = 3;
-    for leader_count in 1..committee.len() {
-        let (mut storage, _) =
-            Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
-        build_dag(&committee, &mut storage, None, 5);
 
-        let mut committer = Committer::new(
-            committee.clone(),
-            storage.block_reader().clone(),
-            Protocol {
-                direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
-                direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
-                anchor_link_size: 1,
-                wave_length: 3,
-                leader_count: NonZeroUsize::new(leader_count).unwrap(),
-                pipeline: false,
-                leader_wait: false,
-                require_crypto: false,
-            },
-        );
+    let (mut storage, _) =
+        Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
+    build_dag(&committee, &mut storage, None, 5);
 
-        let last_committed: Option<(RoundNumber, Authority)> = None;
-        let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
-        tracing::info!("Commit sequence: {sequence:?}");
+    let mut committer = Committer::new(
+        committee.clone(),
+        storage.block_reader().clone(),
+        Protocol {
+            direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
+            direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
+            anchor_link_size: 1,
+            wave_length: 3,
+            leader_count: NonZeroUsize::new(1).unwrap(),
+            pipeline: false,
+            leader_wait: false,
+            require_crypto: false,
+        },
+    );
 
-        assert_eq!(sequence.len(), leader_count);
-        for (i, leader) in sequence.iter().enumerate() {
-            if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) = leader
-            {
-                let leader_round = wave_length;
-                let leader_offset = i as u64;
-                let expected = leader_elector.elect_leader(leader_round + leader_offset);
-                assert_eq!(block.author(), expected);
-            } else {
-                panic!("Expected a committed leader")
-            };
-        }
-    }
+    let last_committed: Option<(RoundNumber, Authority)> = None;
+    let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
+    tracing::info!("Commit sequence: {sequence:?}");
+
+    assert_eq!(sequence.len(), 1);
+    if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
+        sequence[0]
+    {
+        assert_eq!(block.author(), leader_elector.elect_leader(3))
+    } else {
+        panic!("Expected a committed leader")
+    };
 }
 
 /// Ensure idempotent replies.
@@ -64,49 +58,45 @@ fn direct_commit() {
 #[tracing_test::traced_test]
 fn idempotence() {
     let committee = committee(4);
-    for leader_count in 1..committee.len() {
-        let (mut storage, _) =
-            Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
-        build_dag(&committee, &mut storage, None, 5);
 
-        let mut committer = Committer::new(
-            committee.clone(),
-            storage.block_reader().clone(),
-            Protocol {
-                direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
-                direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
-                anchor_link_size: 1,
-                wave_length: 3,
-                leader_count: NonZeroUsize::new(leader_count).unwrap(),
-                pipeline: false,
-                leader_wait: false,
-                require_crypto: false,
-            },
-        );
+    let (mut storage, _) =
+        Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
+    build_dag(&committee, &mut storage, None, 5);
 
-        // Commit one block.
-        let last_committed: Option<(RoundNumber, Authority)> = None;
-        let committed = committer.try_commit(last_committed).collect::<Vec<_>>();
+    let mut committer = Committer::new(
+        committee.clone(),
+        storage.block_reader().clone(),
+        Protocol {
+            direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
+            direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
+            anchor_link_size: 1,
+            wave_length: 3,
+            leader_count: NonZeroUsize::new(1).unwrap(),
+            pipeline: false,
+            leader_wait: false,
+            require_crypto: false,
+        },
+    );
 
-        // Ensure we don't commit it again.
-        let last = committed.into_iter().last().unwrap();
-        let last_committed = Some((last.round(), last.authority()));
-        let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
-        tracing::info!("Commit sequence: {sequence:?}");
-        assert!(sequence.is_empty());
-    }
+    // Commit one block.
+    let last_committed: Option<(RoundNumber, Authority)> = None;
+    let committed = committer.try_commit(last_committed).collect::<Vec<_>>();
+
+    // Ensure we don't commit it again.
+    let max = committed.into_iter().max().unwrap();
+    let last_committed = Some((max.round(), max.authority()));
+    let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
+    tracing::info!("Commit sequence: {sequence:?}");
+    assert!(sequence.is_empty());
 }
 
-/// Commit one by one each wave as the dag progresses in ideal conditions.
+/// Commit one by one each leader as the dag progresses in ideal conditions.
 #[test]
 #[tracing_test::traced_test]
 fn multiple_direct_commit() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let mut last_committed: Option<(RoundNumber, Authority)> = None;
     for n in 1..=10 {
@@ -123,7 +113,7 @@ fn multiple_direct_commit() {
                 direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
                 anchor_link_size: 1,
                 wave_length: 3,
-                leader_count: NonZeroUsize::new(leader_count).unwrap(),
+                leader_count: NonZeroUsize::new(1).unwrap(),
                 pipeline: false,
                 leader_wait: false,
                 require_crypto: false,
@@ -132,72 +122,19 @@ fn multiple_direct_commit() {
 
         let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
         tracing::info!("Commit sequence: {sequence:?}");
-        assert_eq!(sequence.len(), leader_count);
+        assert_eq!(sequence.len(), 1);
 
         let leader_round = n * wave_length;
-        for (i, leader) in sequence.iter().enumerate() {
-            if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) = leader
-            {
-                let leader_offset = i as u64;
-                let expected = leader_elector.elect_leader(leader_round + leader_offset);
-                assert_eq!(block.author(), expected);
-            } else {
-                panic!("Expected a committed leader")
-            };
-        }
-
-        let last = sequence.iter().last().unwrap();
-        last_committed = Some((last.round(), last.authority()));
-    }
-}
-
-/// Commit the leaders of the first wave assuming the very first leader is already committed.
-#[test]
-#[tracing_test::traced_test]
-fn direct_commit_partial_round() {
-    let committee = committee(4);
-    let leader_elector = LeaderElector::new(committee.len());
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
-    let wave_length = 3;
-    let leader_count = strong_quorum as usize;
-
-    let first_leader_round = wave_length;
-    let first_leader = leader_elector.elect_leader(first_leader_round);
-    let last_committed = Some((first_leader_round, first_leader));
-
-    let enough_blocks = 2 * wave_length - 1;
-    let (mut storage, _) =
-        Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
-    build_dag(&committee, &mut storage, None, enough_blocks);
-
-    let mut committer = Committer::new(
-        committee.clone(),
-        storage.block_reader().clone(),
-        Protocol {
-            direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
-            direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
-            anchor_link_size: 1,
-            wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
-            pipeline: false,
-            leader_wait: false,
-            require_crypto: false,
-        },
-    );
-
-    let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
-    tracing::info!("Commit sequence: {sequence:?}");
-    assert_eq!(sequence.len(), leader_count - 1);
-
-    for (i, leader) in sequence.iter().enumerate() {
-        if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) = leader {
-            let leader_offset = (i + 1) % committee.len();
-            let expected = leader_elector.elect_leader(first_leader_round + leader_offset as u64);
-            assert_eq!(block.author(), expected);
+        if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
+            sequence[0]
+        {
+            assert_eq!(block.author(), leader_elector.elect_leader(leader_round));
         } else {
             panic!("Expected a committed leader")
-        };
+        }
+
+        let max = sequence.iter().max().unwrap();
+        last_committed = Some((max.round(), max.authority()));
     }
 }
 
@@ -207,10 +144,7 @@ fn direct_commit_partial_round() {
 fn direct_commit_late_call() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let n = 10;
     let enough_blocks = wave_length * (n + 1) - 1;
@@ -226,7 +160,7 @@ fn direct_commit_late_call() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -237,19 +171,16 @@ fn direct_commit_late_call() {
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
 
-    assert_eq!(sequence.len(), leader_count * n as usize);
-    for (i, leaders) in sequence.chunks(leader_count).enumerate() {
+    assert_eq!(sequence.len(), n as usize);
+    for (i, leader_block) in sequence.iter().enumerate() {
         let leader_round = (i as u64 + 1) * wave_length;
-        for (j, leader) in leaders.iter().enumerate() {
-            if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) = leader
-            {
-                let leader_offset = j as u64;
-                let expected = leader_elector.elect_leader(leader_round + leader_offset);
-                assert_eq!(block.author(), expected);
-            } else {
-                panic!("Expected a committed leader")
-            }
-        }
+        if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) =
+            leader_block
+        {
+            assert_eq!(block.author(), leader_elector.elect_leader(leader_round));
+        } else {
+            panic!("Expected a committed leader")
+        };
     }
 }
 
@@ -258,10 +189,7 @@ fn direct_commit_late_call() {
 #[tracing_test::traced_test]
 fn no_genesis_commit() {
     let committee = committee(4);
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let first_commit_round = 2 * wave_length - 1;
     for r in 0..first_commit_round {
@@ -277,7 +205,7 @@ fn no_genesis_commit() {
                 direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
                 anchor_link_size: 1,
                 wave_length: 3,
-                leader_count: NonZeroUsize::new(leader_count).unwrap(),
+                leader_count: NonZeroUsize::new(1).unwrap(),
                 pipeline: false,
                 leader_wait: false,
                 require_crypto: false,
@@ -291,16 +219,13 @@ fn no_genesis_commit() {
     }
 }
 
-/// We directly skip a leader that has enough blames and commit the others
+/// We directly skip the leader if it is missing.
 #[test]
 #[tracing_test::traced_test]
 fn no_leader() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let (mut storage, _) =
         Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
@@ -309,7 +234,7 @@ fn no_leader() {
     let decision_round_0 = wave_length - 1;
     let references = build_dag(&committee, &mut storage, None, decision_round_0);
 
-    // Add enough blocks to reach the decision round of wave 1 (but without its leader).
+    // Add enough blocks to reach the decision round of the first leader (but without the leader).
     let leader_round_1 = wave_length;
     let leader_1 = leader_elector.elect_leader(leader_round_1);
 
@@ -322,7 +247,7 @@ fn no_leader() {
     let decision_round_1 = 2 * wave_length - 1;
     build_dag(&committee, &mut storage, Some(references), decision_round_1);
 
-    // Ensure the omitted leader is skipped and the others are committed.
+    // Ensure no blocks are committed.
     let mut committer = Committer::new(
         committee.clone(),
         storage.block_reader().clone(),
@@ -331,7 +256,7 @@ fn no_leader() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -342,27 +267,14 @@ fn no_leader() {
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
 
-    assert_eq!(sequence.len(), leader_count);
-    for (i, leader) in sequence.iter().enumerate() {
-        let leader_round = wave_length;
-        let leader_offset = i as u64;
-        let expected_leader = leader_elector.elect_leader(leader_round + leader_offset);
-        if i == 0 {
-            if let LeaderStatus::DirectSkip(leader, round)
-            | LeaderStatus::IndirectSkip(leader, round) = sequence[i]
-            {
-                assert_eq!(leader, expected_leader);
-                assert_eq!(round, leader_round_1);
-            } else {
-                panic!("Expected to directly skip the leader");
-            }
-        } else if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) =
-            leader
-        {
-            assert_eq!(block.author(), expected_leader);
-        } else {
-            panic!("Expected a committed leader")
-        }
+    assert_eq!(sequence.len(), 1);
+    if let LeaderStatus::DirectSkip(leader, round) | LeaderStatus::IndirectSkip(leader, round) =
+        sequence[0]
+    {
+        assert_eq!(leader, leader_1);
+        assert_eq!(round, leader_round_1);
+    } else {
+        panic!("Expected to directly skip the leader");
     }
 }
 
@@ -372,15 +284,12 @@ fn no_leader() {
 fn direct_skip() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
-    let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let (mut storage, _) =
         Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
 
-    // Add enough blocks to reach the first leader of wave 1.
+    // Add enough blocks to reach the first leader.
     let leader_round_1 = wave_length;
     let references_1 = build_dag(&committee, &mut storage, None, leader_round_1);
 
@@ -390,7 +299,7 @@ fn direct_skip() {
         .filter(|x| x.authority != leader_elector.elect_leader(leader_round_1))
         .collect();
 
-    // Add enough blocks to reach the decision round of wave 1.
+    // Add enough blocks to reach the decision round of the first leader.
     let decision_round_1 = 2 * wave_length - 1;
     build_dag(
         &committee,
@@ -399,7 +308,7 @@ fn direct_skip() {
         decision_round_1,
     );
 
-    // Ensure that the first leader of wave 1 is skipped.
+    // Ensure the leader is skipped.
     let mut committer = Committer::new(
         committee.clone(),
         storage.block_reader().clone(),
@@ -408,7 +317,7 @@ fn direct_skip() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -419,27 +328,14 @@ fn direct_skip() {
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
 
-    assert_eq!(sequence.len(), leader_count);
-    for (i, leader) in sequence.iter().enumerate() {
-        let leader_round = wave_length;
-        let leader_offset = i as u64;
-        let expected_leader = leader_elector.elect_leader(leader_round + leader_offset);
-        if i == 0 {
-            if let LeaderStatus::DirectSkip(leader, round)
-            | LeaderStatus::IndirectSkip(leader, round) = sequence[i]
-            {
-                assert_eq!(leader, expected_leader);
-                assert_eq!(round, leader_round_1);
-            } else {
-                panic!("Expected to directly skip the leader");
-            }
-        } else if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) =
-            leader
-        {
-            assert_eq!(block.author(), expected_leader);
-        } else {
-            panic!("Expected a committed leader")
-        }
+    assert_eq!(sequence.len(), 1);
+    if let LeaderStatus::DirectSkip(leader, round) | LeaderStatus::IndirectSkip(leader, round) =
+        sequence[0]
+    {
+        assert_eq!(leader, leader_elector.elect_leader(leader_round_1));
+        assert_eq!(round, leader_round_1);
+    } else {
+        panic!("Expected to directly skip the leader");
     }
 }
 
@@ -453,23 +349,22 @@ fn indirect_commit() {
     let strong_quorum = 2 * total / 3 + 1;
     let one_fault = total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let (mut storage, _) =
         Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
 
-    // Add enough blocks to reach the leaders of wave 1.
+    // Add enough blocks to reach the 1st leader.
     let leader_round_1 = wave_length;
     let references_1 = build_dag(&committee, &mut storage, None, leader_round_1);
 
-    // Filter out the first leader of wave 1.
+    // Filter out that leader.
     let references_without_leader_1: Vec<_> = references_1
         .iter()
         .cloned()
         .filter(|x| x.authority != leader_elector.elect_leader(leader_round_1))
         .collect();
 
-    // Only 2f+1 replicas vote for the that leader.
+    // Only 2f+1 replicas vote for the 1st leader.
     let connections_with_leader_1 = committee
         .authorities()
         .take(strong_quorum as usize)
@@ -486,7 +381,7 @@ fn indirect_commit() {
     let references_without_votes_for_leader_1 =
         build_dag_layer(connections_without_leader_1, &mut storage);
 
-    // Only f+1 replicas certify that leader.
+    // Only f+1 replicas certify the 1st leader.
     let mut references_3 = Vec::new();
 
     let connections_with_votes_for_leader_1 = committee
@@ -514,7 +409,7 @@ fn indirect_commit() {
         &mut storage,
     ));
 
-    // Add enough blocks to decide the leaders of wave 2.
+    // Add enough blocks to decide the 2nd leader.
     let decision_round_3 = 3 * wave_length - 1;
     build_dag(
         &committee,
@@ -532,7 +427,7 @@ fn indirect_commit() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -542,7 +437,7 @@ fn indirect_commit() {
     let last_committed: Option<(RoundNumber, Authority)> = None;
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
-    assert_eq!(sequence.len(), 2 * leader_count);
+    assert_eq!(sequence.len(), 2);
 
     let leader_round = wave_length;
     let leader = leader_elector.elect_leader(leader_round);
@@ -555,26 +450,24 @@ fn indirect_commit() {
     };
 }
 
-/// Commit the leaders of wave 1, skip the first leader of wave 2, and commit the leaders of wave 3.
+/// Commit the first leader, skip the 2nd, and commit the 3rd leader.
 #[test]
 #[tracing_test::traced_test]
 fn indirect_skip() {
     let committee = committee(4);
     let leader_elector = LeaderElector::new(committee.len());
     let total = committee.total_stake();
-    let strong_quorum = 2 * total / 3 + 1;
     let one_fault = total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let (mut storage, _) =
         Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
 
-    // Add enough blocks to reach the leaders of wave 2.
+    // Add enough blocks to reach the 2nd leader.
     let leader_round_2 = 2 * wave_length;
     let references_2 = build_dag(&committee, &mut storage, None, leader_round_2);
 
-    // Filter out the first leader of wave 2.
+    // Filter out that leader.
     let leader_2 = leader_elector.elect_leader(leader_round_2);
     let references_without_leader_2: Vec<_> = references_2
         .iter()
@@ -582,7 +475,7 @@ fn indirect_skip() {
         .filter(|x| x.authority != leader_2)
         .collect();
 
-    // Only f+1 replicas connect to that leader.
+    // Only f+1 replicas connect to the 2nd leader.
     let mut references = Vec::new();
 
     let connections_with_leader_2 = committee
@@ -599,7 +492,7 @@ fn indirect_skip() {
         .collect();
     references.extend(build_dag_layer(connections_without_leader_2, &mut storage));
 
-    // Add enough blocks to reach the decision round of the the third wave.
+    // Add enough blocks to reach the decision round of the 3rd leader.
     let decision_round_3 = 4 * wave_length - 1;
     build_dag(&committee, &mut storage, Some(references), decision_round_3);
 
@@ -612,7 +505,7 @@ fn indirect_skip() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -622,24 +515,23 @@ fn indirect_skip() {
     let last_committed: Option<(RoundNumber, Authority)> = None;
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
-    assert_eq!(sequence.len(), 3 * leader_count);
+    assert_eq!(sequence.len(), 3);
 
-    // Ensure we commit the leaders of wave 1.
-    for (n, status) in sequence.iter().take(leader_count).enumerate() {
-        let leader_round_1 = wave_length;
-        let leader_offset = n as u64;
-        let leader_1 = leader_elector.elect_leader(leader_round_1 + leader_offset);
-        if let LeaderStatus::DirectCommit(block) | LeaderStatus::IndirectCommit(block) = status {
-            assert_eq!(block.author(), leader_1);
-        } else {
-            panic!("Expected a committed leader")
-        };
-    }
+    // Ensure we commit the leader of wave 1.
+    let leader_round_1 = wave_length;
+    let leader_1 = leader_elector.elect_leader(leader_round_1);
+    if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
+        sequence[0]
+    {
+        assert_eq!(block.author(), leader_1);
+    } else {
+        panic!("Expected a committed leader")
+    };
 
-    // Ensure we skip the first leader of wave 2 but commit the other leaders of wave 2.
+    // Ensure we skip the 2nd leader.
     let leader_round_2 = 2 * wave_length;
     if let LeaderStatus::DirectSkip(leader, round) | LeaderStatus::IndirectSkip(leader, round) =
-        sequence[leader_count]
+        sequence[1]
     {
         assert_eq!(leader, leader_2);
         assert_eq!(round, leader_round_2);
@@ -647,42 +539,15 @@ fn indirect_skip() {
         panic!("Expected a skipped leader")
     }
 
-    for n in 0..leader_count {
-        let leader_round_2 = 2 * wave_length;
-        let leader_offset = n as u64;
-        if n == 0 {
-            if let LeaderStatus::DirectSkip(leader, round)
-            | LeaderStatus::IndirectSkip(leader, round) = sequence[leader_count + n]
-            {
-                assert_eq!(leader, leader_2);
-                assert_eq!(round, leader_round_2);
-            } else {
-                panic!("Expected a skipped leader")
-            }
-        } else {
-            let leader_2 = leader_elector.elect_leader(leader_round_2 + leader_offset);
-            if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
-                sequence[leader_count + n]
-            {
-                assert_eq!(block.author(), leader_2);
-            } else {
-                panic!("Expected a committed leader")
-            }
-        }
-    }
-
-    // Ensure we commit the leaders of wave 3.
-    for n in 0..leader_count {
-        let leader_round_3 = 3 * wave_length;
-        let leader_offset = n as u64;
-        let leader_3 = leader_elector.elect_leader(leader_round_3 + leader_offset);
-        if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
-            sequence[2 * leader_count + n]
-        {
-            assert_eq!(block.author(), leader_3);
-        } else {
-            panic!("Expected a committed leader")
-        }
+    // Ensure we commit the 3rd leader.
+    let leader_round_3 = 3 * wave_length;
+    let leader_3 = leader_elector.elect_leader(leader_round_3);
+    if let LeaderStatus::DirectCommit(ref block) | LeaderStatus::IndirectCommit(ref block) =
+        sequence[2]
+    {
+        assert_eq!(block.author(), leader_3);
+    } else {
+        panic!("Expected a committed leader")
     }
 }
 
@@ -695,34 +560,33 @@ fn undecided() {
     let total = committee.total_stake();
     let strong_quorum = 2 * total / 3 + 1;
     let wave_length = 3;
-    let leader_count = strong_quorum as usize;
 
     let (mut storage, _) =
         Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
 
-    // Add enough blocks to reach the leaders of wave 1.
+    // Add enough blocks to reach the first leader.
     let leader_round_1 = wave_length;
     let references_1 = build_dag(&committee, &mut storage, None, leader_round_1);
 
-    // Filter out the first leader of wave 1.
-    let references_1_without_leader: Vec<_> = references_1
+    // Filter out that leader.
+    let references_without_leader_1: Vec<_> = references_1
         .iter()
         .cloned()
         .filter(|x| x.authority != leader_elector.elect_leader(leader_round_1))
         .collect();
 
-    // Create a dag layer where only one authority votes for that leader.
+    // Create a dag layer where only one authority votes for the first leader.
     let mut authorities = committee.authorities();
     let leader_connection = vec![(authorities.next().unwrap(), references_1)];
     let non_leader_connections: Vec<_> = authorities
         .take((strong_quorum - 1) as usize)
-        .map(|authority| (authority, references_1_without_leader.clone()))
+        .map(|authority| (authority, references_without_leader_1.clone()))
         .collect();
 
     let connections = leader_connection.into_iter().chain(non_leader_connections);
     let references = build_dag_layer(connections.collect(), &mut storage);
 
-    // Add enough blocks to reach the decision round of wave 1.
+    // Add enough blocks to reach the decision round of the first leader.
     let decision_round_1 = 2 * wave_length - 1;
     build_dag(&committee, &mut storage, Some(references), decision_round_1);
 
@@ -735,7 +599,7 @@ fn undecided() {
             direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
             anchor_link_size: 1,
             wave_length: 3,
-            leader_count: NonZeroUsize::new(leader_count).unwrap(),
+            leader_count: NonZeroUsize::new(1).unwrap(),
             pipeline: false,
             leader_wait: false,
             require_crypto: false,
@@ -746,4 +610,63 @@ fn undecided() {
     let sequence = committer.try_commit(last_committed).collect::<Vec<_>>();
     tracing::info!("Commit sequence: {sequence:?}");
     assert!(sequence.is_empty());
+}
+
+/// Once a Skip is yielded, the next `try_commit` call seeded with that skip's slot must
+/// not re-yield it. Otherwise `committed_leaders_total{commit_type=*-skip}` would inflate
+/// on every tick that rederives the same trailing skip from the (unchanged) DAG.
+#[test]
+#[tracing_test::traced_test]
+fn trailing_skip_not_re_yielded() {
+    let committee = committee(4);
+    let leader_elector = LeaderElector::new(committee.len());
+    let wave_length = 3;
+
+    let (mut storage, _) =
+        Storage::new_for_tests(Authority::from(0u64), Metrics::new_for_test(0), &committee);
+
+    // Build the same DAG as `direct_skip` so the only decision is one Skip.
+    let leader_round_1 = wave_length;
+    let references_1 = build_dag(&committee, &mut storage, None, leader_round_1);
+    let references_without_leader_1: Vec<_> = references_1
+        .into_iter()
+        .filter(|x| x.authority != leader_elector.elect_leader(leader_round_1))
+        .collect();
+    let decision_round_1 = 2 * wave_length - 1;
+    build_dag(
+        &committee,
+        &mut storage,
+        Some(references_without_leader_1),
+        decision_round_1,
+    );
+
+    let mut committer = Committer::new(
+        committee.clone(),
+        storage.block_reader().clone(),
+        Protocol {
+            direct_commit_quorum: 2 * committee.total_stake() / 3 + 1,
+            direct_skip_quorum: 2 * committee.total_stake() / 3 + 1,
+            anchor_link_size: 1,
+            wave_length: 3,
+            leader_count: NonZeroUsize::new(1).unwrap(),
+            pipeline: false,
+            leader_wait: false,
+            require_crypto: false,
+        },
+    );
+
+    let first = committer.try_commit(None).collect::<Vec<_>>();
+    assert!(
+        matches!(
+            first.last(),
+            Some(LeaderStatus::DirectSkip(..) | LeaderStatus::IndirectSkip(..)),
+        ),
+        "test precondition: last decision must be a Skip, got {first:?}",
+    );
+
+    let seed = first
+        .last()
+        .map(|status| (status.round(), status.authority()));
+    let second = committer.try_commit(seed).collect::<Vec<_>>();
+    assert!(second.is_empty(), "trailing skip re-yielded: {second:?}");
 }
